refactor cli
using multiple entry points
