+++ WrightTools/cli/_tree.py
@@ -1,0 +1,36 @@
+# --- import --------------------------------------------------------------------------------------
+
+
+import click
+import WrightTools as wt
+
+
+# --- define --------------------------------------------------------------------------------------
+
+
+@click.group()
+@click.version_option(wt.__version__)
+def cli():
+    pass
+
+
+@cli.command(name="tree", help="Print a given data tree.")
+@click.argument("path", nargs=1)
+@click.option(
+    "--internal_path", default="/", help="specify a path internal to the file.  Defaults to root"
+)
+@click.option("--depth", "-d", "-L", type=int, default=9, help="Depth to print.")
+@click.option("--verbose", "-v", is_flag=True, default=False, help="Print a more detailed tree.")
+def tree(path, internal_path, depth=9, verbose=False):
+    # open the object
+    obj = wt.open(path)[internal_path]
+
+    if isinstance(obj, wt.Data):
+        obj.print_tree(verbose=verbose)
+    else:
+        obj.print_tree(verbose=verbose, depth=depth)
+
+
+
+if __name__ == "__main__":
+    cli()
